ble client app: Major refactor
The previous version of the code was unable to be run from the terminal (but it worked when run through the PyCharm run command).

A major restructure of the content was made such that we can successfully execute the code from PyCharm or from the terminal.

The new refactor aligns more with best Python project practices.